requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.1
pycryptodome>=3.10.1
python-dotenv>=0.19.0